Added delete of the elements with volume = 0
