+++ cmake-build-debug/CMakeFiles/yellow.dir/CXX.includecache
@@ -13,8 +13,6 @@ tuple
 -
 map
 -
-utility
--
 sstream
 -
 vector
