+++ errors.txt
@@ -1,0 +1,36 @@
+Failed case #1/102: (Wrong answer)
+Неправильный результат в 5 строке
+
+Input:
+AddNewTasks Alice 5
+PerformPersonTasks Alice 5
+PerformPersonTasks Alice 5
+PerformPersonTasks Alice 1
+AddNewTasks Alice 5
+PerformPersonTasks Alice 2
+GetPersonTasksInfo Alice
+PerformPersonTasks Alice 4
+GetPersonTasksInfo Alice
+
+Your output:
+[]
+[{"IN_PROGRESS": 5}, {}]
+[{"TESTING": 5}, {}]
+[{"DONE": 1}, {"TESTING": 4}]
+[]
+[{"IN_PROGRESS": 2, "DONE": 2, "DONE": 1}, {"NEW": 3, "TESTING": 2}]
+{"NEW": 3, "IN_PROGRESS": 2, "TESTING": 2, "DONE": 2, "DONE": 1, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0}
+[{"IN_PROGRESS": 3, "TESTING": 1, "DONE": 1, "DONE": 1, "DONE": 1}, {"IN_PROGRESS": 1, "TESTING": 1, "DONE": 1}]
+{"NEW": 0, "IN_PROGRESS": 4, "TESTING": 2, "DONE": 2, "DONE": 1, "DONE": 1, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0, "DONE": 0}
+
+Correct output:
+[]
+[{"IN_PROGRESS": 5}, {}]
+[{"TESTING": 5}, {}]
+[{"DONE": 1}, {"TESTING": 4}]
+[]
+[{"IN_PROGRESS": 2}, {"NEW": 3, "TESTING": 4}]
+{"NEW": 3, "IN_PROGRESS": 2, "TESTING": 4, "DONE": 1}
+[{"IN_PROGRESS": 3, "TESTING": 1}, {"IN_PROGRESS": 1, "TESTING": 4}]
+{"IN_PROGRESS": 4, "TESTING": 5, "DONE": 1}
+ (Time used: 0.01/1.00, preprocess time used: 0/None, memory used: 44797952/536870912.)
